+++ modules/LifeGame/LifeGame.hh
@@ -1,0 +1,11 @@
+#pragma once
+#ifndef LIFE_GAME_H_
+#define LIFE_GAME_H_
+
+class LifeGame {
+  public:
+    LifeGame();
+    ~LifeGame();
+};
+
+#endif