+++ modules/LifeGame/LifeGameController.hh
@@ -1,0 +1,11 @@
+#pragma once
+#ifndef LIFE_GAME_CONTROLLER_H_
+#define LIFE_GAME_CONTROLLER_H_
+
+class LifeGameController {
+  public:
+    LifeGameController();
+    ~LifeGameController();
+};
+
+#endif