+++ modules/LifeGame/LifeGameControllerWindow.cc
@@ -1,0 +1,15 @@
+/*================================================================
+*  Copyright (C)2020 All rights reserved.
+*  FileName : LifeGameControllerWindow.cc
+*  Author   : elonkou
+*  Email    : elonkou@ktime.cc
+*  Date     : 2020年02月23日 星期日 17时46分19秒
+================================================================*/
+
+#include "LifeGameControllerWindow.hh"
+
+LifeGameControllerWindow::LifeGameControllerWindow() {
+}
+
+LifeGameControllerWindow::~LifeGameControllerWindow() {
+}