+++ modules/LifeGame/LifeGameControllerWindow.hh
@@ -1,0 +1,11 @@
+#pragma once
+#ifndef LIFE_CONTROLLER_WINDOW_H_
+#define LIFE_CONTROLLER_WINDOW_H_
+
+class LifeGameControllerWindow {
+  public:
+    LifeGameControllerWindow();
+    ~LifeGameControllerWindow();
+};
+
+#endif