+++ modules/LifeGame/LifeGameModule.cc
@@ -1,0 +1,15 @@
+/*================================================================
+*  Copyright (C)2020 All rights reserved.
+*  FileName : LifeGameModule.cc
+*  Author   : elonkou
+*  Email    : elonkou@ktime.cc
+*  Date     : 2020年02月23日 星期日 17时57分35秒
+================================================================*/
+
+#include "LifeGameModule.hh"
+
+LifeGameModule::LifeGameModule() {
+}
+
+LifeGameModule::~LifeGameModule() {
+}