+++ modules/LifeGame/LifeGameModule.hh
@@ -1,0 +1,11 @@
+#pragma once
+#ifndef LIFE_GAME_WINDOW_H_
+#define LIFE_GAME_WINDOW_H_
+
+class LifeGameModule {
+  public:
+    LifeGameModule();
+    ~LifeGameModule();
+};
+
+#endif