+++ modules/LifeGame/LifeGameWindow.cc
@@ -1,0 +1,15 @@
+/*================================================================
+*  Copyright (C)2020 All rights reserved.
+*  FileName : LifeGameWindow.cc
+*  Author   : elonkou
+*  Email    : elonkou@ktime.cc
+*  Date     : 2020年02月23日 星期日 17时43分26秒
+================================================================*/
+
+#include "LifeGameWindow.hh"
+
+LifeGameWindow::LifeGameWindow() {
+}
+
+LifeGameWindow::~LifeGameWindow() {
+}